Started Events and Moved ButtonStatus
Started the events by implementing the didEventOccur for the current
Event enum. Still need to test and clarify a couple of the events
based on the diagram in the SRS.

Also moved the ButtonStatus enum to its own file and made it public
so the Events class has access to the enum.